captcha function path changed
